fixup! [#49] Provide error messages for illegal square brackets in URIs
Moved the functions that compose a verbose URI error message to the Xrefcheck.Util module.
